Added a new feature for estimating allometric coefficients from information in the 'raws' table.
right now bash/allom.sh just calls rscripts/AllomAve.R

AllomAve loops over PFTS and stem vs leaf, queries the allometric equations, and fits a power law function to imputed data.

query.allom.data does the actual querying of the database
allom.BayesFit is the Gibbs sampler for the model (fit in R not JAGS because of the need for the data imputation)

Output is a mcmc.list object for the model and a pdf of the mcmc diagnosics

The code is not yet fully functional, a number of features need to be added to handle raw field data on allometries, allometries beyond leaf and stem (height, canopy radius, etc.), the ability to update trait posteriors with the parameter estimates, and the ability to query the previous posterior as the prior. These will be added as redmine feature requests shortly