Fixed issue 775 - consistency between PFTs and the FIA db
Checks are now in place that assure with logger.errors/warnings that:
 -there are not duplicate species in the selected PFTs (error)
 -the selected PFTs cover all species expected by the FIA db (error)
 -the selected PFTs do not have species not expected by FIA (warn)
In any of these cases, a simple lookup table matching species names
to scpds is read (species_names.csv) to output the names of the
trouble species with a nicely-formatted message. Execution is stopped
after the logger.errors. Test cases are enclosed in pft_consistency.R