Working inversion with simplified PROSPECT model.
Entire TAV procedure is independent of inversion, so related variables (tao1/2, rho1/2, x, y) were precalculated and permanently stored in the "prospect4.dat" (tav script now recalcuates them and rewrites the file).

I extracted coefficients from the exponential integral calculations to make the code more readable. The coefficients reside in a CSV file and in the "prospect4.dat" file.

There is a bug in NIMBLE that doesn't allow directly squaring some vectors (e.g. tao90^2). I changed all instances of this to regular multiplications and the code runs fine. There are still weird issues with having the PROSPECT reflectance calculated in a separate file, so for the time being, the full model resides in both scripts.