Reverting changes related to inserting inputs to db
Per comment @mdietze's comment #2 in gh465, I had changed the code to not
automatically add inputs specified by file name (rather than bety id) to
the database. However, this breaks the later step of adding the likelihood to
bety, since likelihoods must be associated with a valid input id. So, for now reverting.