Working basic inversion with model selection
Segfault was caused by the fact that I was declaring "model as a pointer
in "mh_sampler", when it should have just been a "procedure()".

Fixed another issue where "ar" was set to NaN; fixed this by setting
ar=0 at the beginning (NOTE: This is an array definition; all values of
"ar" are set to 0).

Added "-O3" flag to Fortran compiler. This has minimal performance
implications -- approximately 2 second gain at 10,000 iterations (~ 160
second total runtime), so if it causes problems, I can change it back.