Make sure dbfile is unique (gh-494)
Per Rob's direction at gh-494, dbfile.insert now checks first to
see if a file already exists at the requested machine/path/file
location. If so, then the function returns that record's ID or NA,
depending on whether a new 'reuse' argument is T or F, repsectively.

If no file already exists, then a new record is inserted, and its ID is
returned (as before).

I noticed the function header said it returns a data frame of several
relevant values, but actually only the ID was being returned. Updated
this documentation as well.