module/photosynthesis:
fit.A: small fix to not fail when covariate data is absent

plot.photo: major overhaul. This previously was largely just an unfinished cut-and-paste of the code from the Licor.QC plots. Also added the capacity to separate out plots by curve automatically if the model was fit with random effects on. Doesn't do a response curve if <3 points are present (e.g. if the data was all from A-Ci, it won't make an A-Q plot)

vignette:  Basic vignette on running leaf-level analysis. Needs a lot more text/explaination and to make sure it works with the package data (rather than hard-coded path)