Reinstated ability to manually specify dist file for SA/EA
I think this was removed only unintentionally by David.
In playing with this and writing up the fluxcamp demo, I've been thinking that
we may want to expand the functionality to allow *any* prior. or post.distns
file to be used for SA/EA. It's implemented now such that the file has to
exist within the current workflow directory, which was supposed to be a safeguard.
But that precludes fitting distributions in one workflow and them using them in
another (exactly what we wanted for the flux camp PDA demo). Worth discussing.