Rewrite of write.config.xml.ED2 script (#734)
Following the pseudocode and discussion on issue #734.  The script loops
through PFTs defined in `trait.values`. For each PFT:

1. Default values are loaded from the requisite `history.csv` file.
2. Values specified by `trait.values` are converted to ED units, and
   then they replace, by name, the default values set in step 1.
3. Any constant values, defined by the `defaults` argument (if blank,
   this is automatically set to `settings$constants`, which is the
   typical use anyway), replace the values set by steps 1 and 2.
4. These values are written out to an XML file.

**NOTE:** The `defaults` argument, originally the first argument, has been
moved to the third and last argument, and has been set to default to
`settings$constants`.

I also added a simple test script in the `pecan/models/ed/tests` folder
that people can use to check the XML output for a given set of inputs.