Restore old arguments to invert.auto
NOTE: `parallel=FALSE` is currently not implemented. It will throw a
warning and try with `parallel=TRUE` and throw errors if that fails.
Getting serial mode to work again will require a bit more code, but is
not an immediate priority.

A few old `invert.options` components are deprecated -- namely `burnin`
(which is no longer required, since convergence is checked with
`autoburnin=TRUE`), and `do.lsq.first` and `.after` (which are no longer
relevant because the inversion doesn't start over. The `do.lsq` option
is the only relevant one in this series because it controls how the
inversion is initialized).