Update convergence checking mechanism.
Note that with the latest settings, it takes a really long time to
achieve convergence. Not sure if this is a problem with the `coda`
package's Gelman diagnostic calculation, or just an underlying problem
with PROSPECT inversion. Basically, the issue is that
`coda::gelman.diag` indicates convergence MUCH earlier than the
`gelman.plot` function, which is underlying `burnin`. This commit adds
some flexibility about the stringency of convergence checking (for
instance, by setting `use.confidence = FALSE`), but even with these
settings, convergence still takes a long time.