Updates to site.lst. Was originally just trying to avoid this obnoxious warning: "RS-DBI driver warning: (unrecognized PostgreSQL field type geometry (id:16405) in column 17)"
Which arises from querying all columns of sites when only time_zone was needed. But then noticed that the checks on time_zone were not going to do what I think was anticipated. The db column 'time_zone' is type 'text', so even if it contains a number the is.numeric() check fails. Also, currently it contains only text, so I'm not sure what the idea was anyway. The way I've changed it, if someone *does* enter the gmtOffset directly as a number into sites, then this code will convert it to a numeric and return that. Otherwise the offset is looked up in the 'else' clause.

Note that the time_zone column currently contains text descriptions of time zones, which I presume can be mapped directly to gmtOffset somehow. That would be good functionality to implement since the call to GNtimezone is a bit slow. (This fn could also update the db whenever it does look it up, to avoid needing to do so again later)