Almost always you want `library`, not `require`
I’d like to _hurt_ the person who named R’s `require` function, which
is exactly for when a package is NOT required. There are a few
instances of this being used correctly in the PEcAn code, but almost
all can (and should) be changed to `library`—i.e., the code will not
run without that package.