require RPostgreSQL before callng DBI::dbConnect
Not sure if this is the right approach, but I was getting an error (PostgreSQL driver not found, basically) without it. I imagine that RPostgreSQL used to be a "Depends" so it wasn't a problem; now that it's a "Suggest", it seems you need to load it into the namespace before calling DBI functions? using requireNamespace seems to be Wickham's recommendation for within a function...