Bugfixes to Shiny global sensitivity
Replaced `read.ensemble.output` with manual loop to avoid the 'expressions' pitfall. Might be a good idea to add expression back in eventually for even more powerful plotting, but this is just to get this working for now.

Added a lot of `message` statements throughout to assist in debugging. Anything that gets passed to `stderr` will be printed in the Shiny log, which allows for debugging Shiny runs in the web interface.