Convert make test to make check for CI
`make test` calls `devtoos::test`; `make check` calls `devtools::check`. 

`devtools::test` and `make test` are useful for development, but `devtools::check`, which calls R CMD Check, is more appropriate for continuous integration (according to devtools author [Hadley Wickham's text](http://r-pkgs.had.co.nz/check.html#check). `check`   integrates with the standard protocol for R package testing.

One side effect of using `test` is that it only calls files in `tests/testthat/test*.R` files. `check` calls `tests/testthat.R` which then runs `test_check` and checks the package. If we only use `test` for CI (which I don't think we should) the `testthat.R` files should be deprecated.

A benefit of using `check` is that it performs a standard suite of tests for R packages, and will ensure that the package itself is valid (not just the functions that it contains).

A downside is longer times for Travis CI. But ... if time is an issue perhaps we should parse `git diff` in make and only test the recent changes.