PEcAnRTM: Initialisze RNG on parallel cluster
Otherwise, chains start with same seed and can end up identical. This
perturbs the seed on each chain, forcing them to be different.

For some reason, this has never been a problem for PROSPECT inversions
but is a problem for EDR. Either way, this fix shouldn't break anything,
and will enable EDR to run properly.