open.tunnel: R function to open ssh tunnel. NOTE: location of ssh tunnel script in web dir currently hardcoded. Function needs more testing
PalEON regional met script: addition of code to set up run of extract.nc on cluster and code that pulls files back after running the extraction remotely. There's also a script on geo that I need to pull back and commit that does the extraction via the queue. Had thought about generalizing existing PEcAn multisite papply function to run in parrallel, but to the best of my knowledge our remote met function executions are not run on the cluster (just the head node) so wouldn't have saved much time (and I didn't have time to generalize that too). But a useful idea to pursue if we continue to do pseudoregional runs, since input processing will be nontrivial compute.