Initial De-biasing scripts
Translation of PalEON debiasing (spatial downscaling & bias-correction)
workflow into scripts that can be used throughout Pecan!

align.met function takes a training dataset and makes the dataset to be
downscaled on the same temporal resolution.  This *should* work for
both this bias correction step & the temporal downscaling.

debias.met.regression is a beast of a function that does the bias
correction and attempts to take into account the covariance among
parameters by doing variables in a specific order and then adding
predictors/complexity as we go along.  There are special cases for if
we’re working with two empirical(-ish) datasets where we can pair the
anomalies as well as some additional flags for CRUNCEP.  There are some
improvements that can/will be made (e.g. possibly transforming
zero-truncated variables when used as predictors & not just responses),
but this is what I have so far.