Fix setting origin of time vector
For most cases, p1000 scenarios are referenced off of 850 in all files
and historical is referenced off of 1850.  MPI-ESM-P seems to be an
exception an is referenced off of 850 in all cases.  Could probably
find a way to soft code this, but I think I tried it once before and
found had trouble.