Add code for making the runmask custom to the PEcAn run and...
...moves site selection up to the pecan.xml file (settings object).

NOTE: this is really close to working but having problems casting
data correctly to write to the netcdf file!!!

This will handle the case where the default dvmdostem sample input DATA
has a run-mask with too many, or the wrong pixels activated. The
pixel to activate in the run-mask is exposed in the pecan.xml file
using the <dvmdostem_pixel_X> and <dvmdostem_pixel_Y> tags.

This is important because we don't always want to run pixel 0,0. For example
pixel 0,0 in the DATA/SouthBarrow sample dataset is an ocean pixel.