Add check for the correct pixel/cmt match.
THis is probably unique to dvmdostem, but basically because our default
input dataset is a grid, and we only want to run one pixel from the grid,
it is possible to select a pixel that has a different vegetation (community)
type than the vegetation (community) type that is associated with the PFT
we are running for. When this happens, the incorrect parameters are written
out to the updated parameter files, and the run fails. This is an attempt
to check for this problem up-front. But I am not sure how to correctly
exit a pecan run.