Creating new function bety2pecan for converting bety variables to pecan standard
We may want to move away from using a lookup table to do these conversions and it will only involve changing this function instead of query.format.vars
This is also where I am putting the Ameriflux hack for FC and NEE mapping to NEE in pecan standard. This is only a temporary fix.