DOC: Add documentation for modifying Bety schema
If this exists in the Bety docs already, I couldn't find it. The Bety
docs may be a better place for this -- if so, let me know where to put
it and I'll happily stick it there. Though, for what it's worth, becase
PEcAn and Bety are so tightly integrated, I think it's a good idea to
merge the documentation into a single document.