Add new loop for working with multiple PFTs.
Previously we had hard-coded the trait value lookup to simply look
at the first PFT specified in the settings dictionary. If the user
had specified more PFTs in the xml file, then they were ignored.

With this change, we loop over all the PFTs in the settings object
and update the parameters in our (temporary) json objects accordingly.