Add reduced output_spec that is specific to PEcAn.
This will help reduce the volume of outputs that are generated when
PEcAn is used to make hundreds of dvmdostem runs. This commit adds
a custom output_spec file to the inst folder, and write configs copys
that file into the the run directory for each run that is carried out.

In the future we could add different output_specs to the inst folder if
there is a need to have different "suites" of output variables.