refactor: more control for user specified output_spec.csv files
This allows for users to specify an output spec that may be anywhere
on the filesystem. Also copies output_spec to run folder and retains name
which makes it easier to figure out post-run what variables were turned on.

Also adds another output_spec file with more variables turned on. Use with
caution - may result in a lot of output!