UPDATE wrapper to conform to dvmdostem commit 892154aee4462fa1e
The dvmdostem codebase changed a default parameter name, and thus, this
wrapper code needs this update so it can find the correct parameter when
parsing dvmdostem parameter files.

dvmdostem does not have a versioning scheme in place (i.e. numbers like
v1.0, etc) so instead we are refering here to the commit SHA in the
dvmdostem repository that necessitates this change.