Finish wiring things up so that output variables are controlled by...
specifying PEcAn names in the xml file. Then the "depends" list
is used to figure out which dvmdostem variables need to be turned
on. Appropriate output spec files are generated based on the
depends lists and dvmdostem generates the correct files (outputs)
to support the requested PEcAn variables.

There is a mapping held in a separate file for tracking how the PEcAn and
dvmdostem variables relate to eachother.

This needs some refactoring, as right now the variable names are confusing.