Implement dvmdostem calibration parameters in write.config.dvmdostem.R
To use this, in bety, create new PFTs that are the "calibraton variants" of
the PFTs you'd like to calibrate for. The "variants" will have a .cal
suffix on their bety name, i.e. CMT06-Sphagnum.cal and will only these
calibration parameters set as related priors.

First step toward using PEcAn to calibrate dvmdostem. Aiming to perform
a sensitivity analysis across the calibration parameters to
see which parameters are most important.

One issue with this implementation is that the 5 soil parameters are
included with each PFT, which will result in unnecessary runs.