[WIP] Remove a lot of text from Git workflow
I was motivated to just remove the "Basic Workflow" section that starts with "This workflow is for educational purposes only." Then I continued to remove content that I thought was not necessary. A lot of this content dates from previous decades before the carpentries were widespread and before git literacy was common among young scientists.

I tried to begin removing a lot of content that can easily be found on google. 

This section needs a lot of work to be useful - we should make it clear (and consistent) if forking is expected of everyone, or just preferred, etc. Currently there is a focus on branching with a few allusions to forking. And many references to cloning from pecanproject/pecan and branching imply forking isn't the first priority (or when it is). NB: I have a problem with forks because it makes it harder to edit text files in the github web interface.