Fix temporal trends in tair
a logically error was removing temporal trends from temperature rather than adding it in.  To get the temporal trend back with uncertainty, we essentially want the residual trend from the simulated tdistribution.  This means removing the mean trend from the air temp anomalies and then adding back in a randomly distributed trend based on the Rbeta distribution.  For a single-member ensemble, we could just work with the temp anomalies since we're workign with the model best-estimate coefficients rather than a simulated distribution, but it makes for more consistent code and logic if we keep it there.