API: Only set model ID, not revision or type
A model ID unambiguously determines model name, type, revision, etc., so
additional information on this in the XML is redundant. Moreover, the
query here was inserting this information _incorrectly_ ---
`models.model_name` is technically _not_ the same as
`model$type` (though, in practice, they often are). This was causing the
ED2 integration tests to fail (because they were looking for a
nonexistent model type `ED2.2`).

The "correct" approach is for users to identify their target
model ID by using API queries and then pass that to the API. Note that
this is already how the `rpecanapi::submit.workflow` function works (it
takes a `model_id` as an argument, not `model_name`, `revision`, etc.).