ED: met2model overhaul
The motivation for this was that the previous version didn't allow ED2
to generate met inputs for incomplete years.  I don't really know why we
previously didn't allow this. I would have liked to implement a less
disruptive fix, but I couldn't understand what the old code was doing.
So I reorganized it in a way that will definitely work for partial
years (I tested!) and, IMHO, is a bit clearer. I did my best to add
explanatory comments for any code that wasn't immediately obvious.