Makefile: install with remotes::install_local instead of devtools::install
devtools::install insists in checking for updates to all dependencies every call,
even when `upgrade = FALSE`. Across all packages this adds up to a lot of repeated
checks, and in combination with our separate dependency check it's enough to exceed
GitHub APi request rate limits during CI builds.

remotes::install_local appears to not check dependencies at all when dependencies=FALSE,
which is what we want for this case.